requirements.txt
pillow==11.3.0
streamlit==1.50.0